Rename ProviderBasedPKIUtility interface.
For compatibility reasons, I originally created the interface for PKI
operations and named it ProviderBasedPKIUtility and then had PKIUtility
implement it.  However, these names are backwards from what one would
expect: PKIUtility sounds like an interface and ProviderBasedPKIUtility
sounds like an implementation.  This patch corrects the confusion in
naming.
